test: move SSR module error test to ssr-vue
