+++ playground/ssr-vue/__tests__/fixtures/ssrModuleLoader-bad.js
