+++ playground/ssr-vue/__tests__/ssr-vue.spec.ts
@@ -1,4 +1,5 @@
 import { resolve } from 'node:path'
+import { fileURLToPath } from 'node:url'
 import fetch from 'node-fetch'
 import { port } from './serve'
 import {
@@ -7,7 +8,8 @@ import {
   getColor,
   isBuild,
   page,
-  untilUpdated
+  untilUpdated,
+  viteServer
 } from '~utils'
 
 const url = `http://localhost:${port}/test/`
@@ -202,3 +204,33 @@ test.runIf(isBuild)('dynamic css file should be preloaded', async () => {
     expect(homeHtml).toMatch(file)
   }
 })
+
+test.runIf(!isBuild)(
+  'always throw error when evaluating an wrong SSR module',
+  async () => {
+    const __filename = fileURLToPath(import.meta.url)
+    const badjs = resolve(__filename, '../fixtures/ssrModuleLoader-bad.js')
+    const THROW_MESSAGE = 'it is an expected error'
+
+    const spy = vi.spyOn(console, 'error').mockImplementation(() => {})
+    const expectedErrors = []
+    for (const _ of [0, 1]) {
+      try {
+        console.log(viteServer)
+        await viteServer.ssrLoadModule(badjs, { fixStacktrace: true })
+      } catch (e) {
+        expectedErrors.push(e)
+      }
+    }
+    expect(expectedErrors).toHaveLength(2)
+    expect(expectedErrors[0]).toBe(expectedErrors[1])
+    expectedErrors.forEach((error) => {
+      expect(error?.message).toContain(THROW_MESSAGE)
+    })
+    expect(spy).toBeCalledTimes(1)
+    const [firstParameter] = spy.mock.calls[0]
+    expect(firstParameter).toContain('Error when evaluating SSR module')
+    expect(firstParameter).toContain(THROW_MESSAGE)
+    spy.mockClear()
+  }
+)
